fix(dp): use u128 for pattern counts to prevent overflow with n ≥ 21
For n = 21 (e.g. `grid 4x4 -f 5`), the number of valid patterns of maximum length can reach 21! ≈ 5.1 × 10¹⁹, which exceeds u64::MAX (≈ 1.8 × 10¹⁹). The DP table itself is unaffected — individual cells are bounded by (n-1)! = 20! < u64::MAX — so only the `counts` accumulator
needed widening to u128.

Adds a regression test (ignored by default, ~350 MB) that verifies `grid 4x4 -f 5` produces counts[21] > u64::MAX without panicking.

+++ src/cli.rs
@@ -139,7 +139,7 @@ fn run_pipeline(grid: &GridDefinition, min_length: usize, max_length: usize, qui
     let counts = count_patterns_dp(n, &blocks, max_length);
     let elapsed = t1.elapsed();
 
-    let total: u64 = counts[min_length..=max_length].iter().sum();
+    let total: u128 = counts[min_length..=max_length].iter().sum();
     let mut lines: Vec<String> = Vec::new();
     for (k, c) in counts.iter().enumerate().skip(min_length) {
         if *c > 0 {
